Persistindo dados da tela de Consulta. E algumas alterações e implementações básicas.

+++ GerenciamentoDeClinica/telamedico/TelaPesquisarMedico.cs
@@ -6,6 +6,7 @@ using System.ComponentModel;
 using System.Data;
 using System.Drawing;
 using System.Linq;
+using System.Net;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -16,6 +17,7 @@ namespace GerenciamentoDeClinica.telamedico
 {
     public partial class TelaPesquisarMedico : Form, IConsistenciaDados
     {
+        private const string ERROR_WEBSERVICE = @"Erro de conexão o servidor.";
         private Thread _threadSalvarDados;
         private string _savedPesquisar = "";
         private PesquisarMedico _pesquisarMedico;
@@ -50,11 +52,12 @@ namespace GerenciamentoDeClinica.telamedico
                     listMedicos.Items[_pesquisarMedico.LinhaSelecionada.Value].Selected = true;
 
                 CarregarEditar(_pesquisarMedico.Medico);
-            } else
+            }
+            else
             {
                 _pesquisarMedico = new PesquisarMedico();
             }
-            
+
             _threadSalvarDados = new Thread(SalvarDados);
             _threadSalvarDados.Start();
         }
@@ -100,25 +103,13 @@ namespace GerenciamentoDeClinica.telamedico
             txtCidade.Enabled = false;
             comboUF.Enabled = false;
 
-            txtNome.Clear();
-            maskedCPF.Clear();
-            txtRG.Clear();
-            maskedCell.Clear();
-            txtCRM.Clear();
+            ClearTextBoxs();
             comboEspecialidade.SelectedIndex = 0;
             dateTimeDtNasc.Value = DateTime.Now;
-            txtEmail.Clear();
             rbSolteiro.Checked = false;
             rbCasado.Checked = false;
             rbViuvo.Checked = false;
-            maskedCEP.Clear();
-            txtLogradouro.Clear();
-            txtNumero.Clear();
-            txtComplemento.Clear();
-            txtBairro.Clear();
-            txtCidade.Clear();
             comboUF.SelectedIndex = 0;
-            txtPais.Clear();
             listMedicos.SelectedItems.Clear();
         }
 
@@ -138,10 +129,16 @@ namespace GerenciamentoDeClinica.telamedico
 
                 DisableEditar();
             }
+            catch (WebException)
+            {
+                MessageBox.Show(this, ERROR_WEBSERVICE, Application.ProductName, MessageBoxButtons.OK,
+                    MessageBoxIcon.Error);
+            }
             catch (Exception ex)
             {
-                MessageBox.Show(this, ex.Message);
+                MessageBox.Show(this, ex.Message, Application.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
+
         }
 
         private void btnAtualizar_Click(object sender, EventArgs e)
@@ -162,9 +159,13 @@ namespace GerenciamentoDeClinica.telamedico
 
                     DisableEditar();
                 }
+                catch (WebException)
+                {
+                    MessageBox.Show(ERROR_WEBSERVICE);
+                }
                 catch (Exception ex)
                 {
-                    MessageBox.Show(this, ex.Message);
+                    MessageBox.Show(this, ex.Message, Application.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
             }
         }
@@ -175,16 +176,31 @@ namespace GerenciamentoDeClinica.telamedico
             {
                 try
                 {
-                    ClinicaService service = new ClinicaService();
-                    service.RemoverMedico(_pesquisarMedico.MedicosSalvos[_pesquisarMedico.LinhaSelecionada.Value]);
-                    MessageBox.Show(@"Médico atualizado com sucesso!");
-                    _pesquisarMedico.MedicosSalvos.RemoveAt(_pesquisarMedico.LinhaSelecionada.Value);
-                    listMedicos.Items.RemoveAt(_pesquisarMedico.LinhaSelecionada.Value);
-                    DisableEditar();
+
+                    var repost = MessageBox.Show(@"Deseja remover o paciente?", @"Confirmação", MessageBoxButtons.YesNo);
+
+                    if (repost == DialogResult.Yes)
+                    {
+                        ClinicaService service = new ClinicaService();
+                        service.RemoverMedico(_pesquisarMedico.MedicosSalvos[_pesquisarMedico.LinhaSelecionada.Value]);
+                        MessageBox.Show(@"Médico removido com sucesso!");
+                        _pesquisarMedico.MedicosSalvos.RemoveAt(_pesquisarMedico.LinhaSelecionada.Value);
+                        listMedicos.Items.RemoveAt(_pesquisarMedico.LinhaSelecionada.Value);
+                        DisableEditar();
+                    }
+                    else
+                    {
+                        DisableEditar();
+                        txtPesqNome.Focus();
+                    }
+                }
+                catch (WebException)
+                {
+                    MessageBox.Show(ERROR_WEBSERVICE);
                 }
                 catch (Exception ex)
                 {
-                    MessageBox.Show(ex.Message);
+                    MessageBox.Show(this, ex.Message, Application.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
             }
         }
@@ -402,6 +418,13 @@ namespace GerenciamentoDeClinica.telamedico
                 Email = txtEmail.Text,
                 Estado_Civil = GetEstadoCivil()
             };
+        }
+
+        void ClearTextBoxs()
+        {
+            Controls.OfType<TextBox>().ToList().ForEach(t => t.Clear());
+            Controls.OfType<MaskedTextBox>().ToList().ForEach(t => t.Clear());
+
         }
 
         private void TelaPesquisarMedico_FormClosing(object sender, FormClosingEventArgs e)
